fix(home): switch parallax-scroll with a normal scroll container to fix layout issue
